Finished MasterToken Least Privilege
This implementation of least privilege separates the usage of a working token and a masterToken. When logging into the groupserver, the user will be asked to select a subset of groups that they would like to work in. The groupserver will then issue them a token that contains only that subset of groups.

This will increase security by not requiring high-privilege tokens to be sent over the network as often as the least-privlege workingtokens.

It also provides convenience to the user through allowing them to select a subset of groups to work under.