File Server Authentication (untested)
Performs file server authentication, in two messages between the file server and the file client. 64-bit random challenges, encrypted with public keys of the two parties. The file server is in charge of generating the AES session key.

This doesn't do anything with the tokens yet, their signatures are verified on a case-by-case basis, separately from the authentication process.